[PROPERTIES] Renamed classes and moved to another new package
