+++ gmail-gateway-api/src/main/java/com/botcamp/gmail_gateway_api/config/properties/DataSourceConfigProperties.java
@@ -1,4 +1,4 @@
-package com.botcamp.gmail_gateway_api.config;
+package com.botcamp.gmail_gateway_api.config.properties;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
