+++ gmail-gateway-api/src/main/java/com/botcamp/gmail_gateway_api/config/properties/GmailUserConfigProperties.java
@@ -1,4 +1,4 @@
-package com.botcamp.gmail_gateway_api.config;
+package com.botcamp.gmail_gateway_api.config.properties;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -7,7 +7,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @ConfigurationProperties(prefix = "gmail.api.user")
 @Data
 @NoArgsConstructor
-public class GmailUserConfig {
+public class GmailUserConfigProperties {
 
     private String email;
 }
